feat: enable low-risk linters (unconvert, wastedassign, usestdlibvars) (#23)
* feat: enable low-risk linters (unconvert, wastedassign, usestdlibvars)

Phase 1 of linting improvements - enable linters that catch real bugs
and improve code quality with minimal false positives.

Changes:
- Enable unconvert: Remove unnecessary type conversions (123 fixes)
- Enable wastedassign: Fix unused variable assignments (3 fixes)
- Enable usestdlibvars: Use stdlib constants (98 fixes)
  - HTTP status codes: use http.StatusOK, http.StatusNotFound, etc.
  - HTTP methods: use http.MethodGet, http.MethodPost, etc.
  - Time layouts: use time.DateTime, time.DateOnly, time.TimeOnly

Total: 224 linting issues resolved across 72 files

All tests pass. No regressions introduced.

* fix: use copy() instead of loop for slice assignment (staticcheck S1001)

+++ internal/adapters/v0_0_41/standalone_adapter.go
@@ -419,7 +419,7 @@ func (a *StandaloneAdapter) GetTRES(ctx context.Context) (*types.TRESList, error
 			tres.Name = *apiTres.Name
 		}
 		if apiTres.Count != nil {
-			tres.Count = int64(*apiTres.Count)
+			tres.Count = *apiTres.Count
 		}
 
 		tresList = append(tresList, tres)
@@ -450,7 +450,7 @@ func (a *StandaloneAdapter) CreateTRES(ctx context.Context, req *types.CreateTRE
 	}
 
 	if req.Count > 0 {
-		count := int64(req.Count)
+		count := req.Count
 		requestTres.Count = &count
 	}
 
@@ -487,7 +487,7 @@ func (a *StandaloneAdapter) CreateTRES(ctx context.Context, req *types.CreateTRE
 	}
 
 	if req.Count > 0 {
-		tres.Count = int64(req.Count)
+		tres.Count = req.Count
 	}
 
 	return tres, nil
