+++ internal/adapters/v0_0_42/reservation_adapter.go
@@ -215,7 +215,7 @@ func (a *ReservationAdapter) convertAPIReservationToCommon(apiReservation api.V0
 	}
 
 	if apiReservation.NodeCount != nil {
-		reservation.NodeCount = int32(*apiReservation.NodeCount)
+		reservation.NodeCount = *apiReservation.NodeCount
 	}
 
 	if apiReservation.Users != nil {
