+++ internal/adapters/v0_0_43/standalone_adapter.go
@@ -429,7 +429,7 @@ func (a *StandaloneAdapter) GetTRES(ctx context.Context) (*types.TRESList, error
 			tres.Name = *apiTres.Name
 		}
 		if apiTres.Count != nil {
-			tres.Count = int64(*apiTres.Count)
+			tres.Count = *apiTres.Count
 		}
 
 		tresList = append(tresList, tres)
@@ -458,7 +458,7 @@ func (a *StandaloneAdapter) CreateTRES(ctx context.Context, req *types.CreateTRE
 	}
 
 	if req.Count > 0 {
-		count := int64(req.Count)
+		count := req.Count
 		apiReq.TRES[0].Count = &count
 	}
 
@@ -482,7 +482,7 @@ func (a *StandaloneAdapter) CreateTRES(ctx context.Context, req *types.CreateTRE
 	}
 
 	if req.Count > 0 {
-		tres.Count = int64(req.Count)
+		tres.Count = req.Count
 	}
 
 	return tres, nil
